+++ tests/navdata/test_navdata.py
@@ -5,8 +5,6 @@
 __authors__ = "A. Kanhere, D. Knowles"
 __date__ = "30 Apr 2022"
 
-
-import os
 import pathlib
 import itertools
 
@@ -16,170 +14,6 @@ import pandas as pd
 from pytest_lazyfixture import lazy_fixture
 
 from gnss_lib_py.navdata.navdata import NavData
-
-def fixture_csv_path(csv_filepath):
-    """Location of measurements for unit test
-
-    Returns
-    -------
-    root_path : string
-        Folder location containing measurements
-    """
-    root_path = os.path.dirname(
-                os.path.dirname(
-                os.path.dirname(
-                os.path.realpath(__file__))))
-    root_path = os.path.join(root_path, 'data/unit_test/navdata')
-
-    csv_path = os.path.join(root_path, csv_filepath)
-
-    return csv_path
-
-@pytest.fixture(name="csv_simple")
-def fixture_csv_simple():
-    """csv with simple format.
-
-    """
-    return fixture_csv_path("navdata_test_simple.csv")
-
-@pytest.fixture(name="csv_headless")
-def fixture_csv_headless():
-    """csv without column names.
-
-    """
-    return fixture_csv_path("navdata_test_headless.csv")
-
-@pytest.fixture(name="csv_missing")
-def fixture_csv_missing():
-    """csv with missing entries.
-
-    """
-    return fixture_csv_path("navdata_test_missing.csv")
-
-@pytest.fixture(name="csv_mixed")
-def fixture_csv_mixed():
-    """csv with mixed data types.
-
-    """
-    return fixture_csv_path("navdata_test_mixed.csv")
-
-@pytest.fixture(name="csv_inf")
-def fixture_csv_inf():
-    """csv with infinity values in numeric columns.
-
-    """
-    return fixture_csv_path("navdata_test_inf.csv")
-
-@pytest.fixture(name="csv_nan")
-def fixture_csv_nan():
-    """csv with NaN values in columns.
-
-    """
-    return fixture_csv_path("navdata_test_nan.csv")
-
-@pytest.fixture(name="csv_int_first")
-def fixture_csv_int_first():
-    """csv where first column are integers.
-
-    """
-    return fixture_csv_path("navdata_test_int_first.csv")
-
-@pytest.fixture(name="csv_only_header")
-def fixture_csv_only_header():
-    """csv where there's no data, only columns.
-
-    """
-    return fixture_csv_path("navdata_only_header.csv")
-
-@pytest.fixture(name="csv_dtypes")
-def fixture_csv_dtypes():
-    """csv made up of different data types.
-
-    """
-    return fixture_csv_path("navdata_test_dtypes.csv")
-
-def load_test_dataframe(csv_filepath, header="infer"):
-    """Create dataframe test fixture.
-
-    """
-
-    data = pd.read_csv(csv_filepath, header=header)
-
-    return data
-
-@pytest.fixture(name='df_simple')
-def fixture_df_simple(csv_simple):
-    """df with simple format.
-
-    """
-    return load_test_dataframe(csv_simple)
-
-@pytest.fixture(name='df_headless')
-def fixture_df_headless(csv_headless):
-    """df without column names.
-
-    """
-    return load_test_dataframe(csv_headless,None)
-
-@pytest.fixture(name='df_missing')
-def fixture_df_missing(csv_missing):
-    """df with missing entries.
-
-    """
-    return load_test_dataframe(csv_missing)
-
-@pytest.fixture(name='df_mixed')
-def fixture_df_mixed(csv_mixed):
-    """df with mixed data types.
-
-    """
-    return load_test_dataframe(csv_mixed)
-
-@pytest.fixture(name='df_inf')
-def fixture_df_inf(csv_inf):
-    """df with infinity values in numeric columns.
-
-    """
-    return load_test_dataframe(csv_inf)
-
-@pytest.fixture(name='df_nan')
-def fixture_df_nan(csv_nan):
-    """df with NaN values in columns.
-
-    """
-    return load_test_dataframe(csv_nan)
-
-@pytest.fixture(name='df_int_first')
-def fixture_df_int_first(csv_int_first):
-    """df where first column are integers.
-
-    """
-    return load_test_dataframe(csv_int_first)
-
-@pytest.fixture(name='df_only_header')
-def fixture_df_only_header(csv_only_header):
-    """df where only headers given and no data.
-
-    """
-    return load_test_dataframe(csv_only_header)
-
-@pytest.fixture(name="data")
-def load_test_navdata(df_simple):
-    """Creates a NavData instance from df_simple.
-
-    """
-    return NavData(pandas_df=df_simple)
-
-@pytest.fixture(name="numpy_array")
-def create_numpy_array():
-    """Create np.ndarray test fixture.
-    """
-    test_array = np.array([[1,2,3,4,5,6],
-                            [0.5,0.6,0.7,0.8,-0.001,-0.3],
-                            [-3.0,-1.2,-100.,-2.7,-30.,-5],
-                            [-543,-234,-986,-123,843,1000],
-                            ])
-    return test_array
 
 def test_init_blank():
     """Test initializing blank NavData class
@@ -1214,248 +1048,6 @@ def test_wrong_init_set(row_idx):
     with pytest.raises(KeyError):
         empty_data[row_idx] = np.zeros([1, 6])
 
-@pytest.fixture(name='add_array')
-def fixture_add_array():
-    """Array added as additional timesteps to NavData from np.ndarray
-
-    Returns
-    -------
-    add_array : np.ndarray
-        Array that will be added to NavData
-    """
-    add_array = np.hstack((10*np.ones([4,1]), 11*np.ones([4,1])))
-    return add_array
-
-
-@pytest.fixture(name='add_df')
-def fixture_add_dataframe():
-    """Pandas DataFrame to be added as additional timesteps to NavData
-
-    Returns
-    -------
-    add_df : pd.DataFrame
-        Dataframe that will be added to NavData
-    """
-    add_data = {'names': np.asarray(['beta', 'alpha'], dtype=object),
-                'integers': np.asarray([-2, 45], dtype=np.int64),
-                'floats': np.asarray([1.4, 1.5869]),
-                'strings': np.asarray(['glonass', 'beidou'], dtype=object)}
-    add_df = pd.DataFrame(data=add_data)
-    return add_df
-
-
-def test_add_numpy(numpy_array, add_array):
-    """Test addition of a numpy array to NavData
-
-    Parameters
-    ----------
-    numpy_array : np.ndarray
-        Array with which NavData instance is initialized
-    add_array : np.ndarray
-        Array to add to NavData
-    """
-    data = NavData(numpy_array=numpy_array)
-    data.concat(NavData(numpy_array=add_array),axis=1,inplace=True)
-    new_col_num = np.shape(add_array)[1]
-    np.testing.assert_array_equal(data[:, -new_col_num:], add_array)
-
-
-def test_add_numpy_1d():
-    """Test addition of a 1D numpy array to NavData with single row
-    """
-    data = NavData(numpy_array=np.zeros([1,6]))
-    data.concat(NavData(numpy_array=np.ones(8)),axis=1, inplace=True)
-    np.testing.assert_array_equal(data[0, :], np.hstack((np.zeros(6),
-                                  np.ones(8))))
-
-    # test adding to empty NavData
-    data_empty = NavData()
-    data_empty.concat(NavData(numpy_array=np.ones((8,8))),axis=1,
-                      inplace=True)
-    np.testing.assert_array_equal(data_empty[:,:],np.ones((8,8)))
-
-def test_add_csv(df_simple, csv_simple):
-    """Test adding a csv.
-
-    """
-    # Create and add to NavData
-    data = NavData(csv_path=csv_simple)
-    data.concat(NavData(csv_path=csv_simple),axis=1,inplace=True)
-    data_df = data.pandas_df()
-    # Set up dataframe for comparison
-    df_types = {'names': object, 'integers': np.int64,
-                'floats': np.float64, 'strings': object}
-    expected_df = pd.concat((df_simple,df_simple)).reset_index(drop=True)
-    expected_df = expected_df.astype(df_types)
-    pd.testing.assert_frame_equal(data_df.sort_index(axis=1),
-                                  expected_df.sort_index(axis=1),
-                                  check_index_type=False)
-
-    # test adding to empty NavData
-    data_empty = NavData()
-    data_empty.concat(NavData(csv_path=csv_simple),axis=1,inplace=True)
-    pd.testing.assert_frame_equal(data_empty.pandas_df().sort_index(axis=1),
-                                  df_simple.astype(df_types).sort_index(axis=1),
-                                  check_index_type=False)
-
-def test_add_pandas_df(df_simple, add_df):
-    """Test addition of a pd.DataFrame to NavData
-
-    Parameters
-    ----------
-    df_simple : pd.DataFrame
-        pd.DataFrame to initialize NavData with
-    add_df : pd.DataFrame
-        pd.DataFrame to add to NavData
-    """
-    data = NavData(pandas_df=df_simple)
-    data.concat(NavData(pandas_df=add_df),axis=1,inplace=True)
-    new_df = data.pandas_df()
-    add_row_num = add_df.shape[0]
-    subset_df = new_df.iloc[-add_row_num:, :].reset_index(drop=True)
-    pd.testing.assert_frame_equal(subset_df.sort_index(axis=1),
-                                  add_df.sort_index(axis=1),
-                                  check_index_type=False)
-
-    # test adding to empty NavData
-    data_empty = NavData()
-    data_empty.concat(NavData(pandas_df=add_df),axis=1,inplace=True)
-    pd.testing.assert_frame_equal(add_df.sort_index(axis=1),
-                                  data_empty.pandas_df().sort_index(axis=1),
-                                  check_index_type=False)
-
-def test_concat(df_simple):
-    """Test concat functionaltiy.
-
-    Parameters
-    ----------
-    df_simple : pd.DataFrame
-        Simple pd.DataFrame with which to initialize NavData.
-
-    """
-
-    navdata_1 = NavData(pandas_df=df_simple)
-    navdata_2 = navdata_1.copy()
-    navdata_2.rename(mapper={"floats": "decimals", "names": "words"},
-                    inplace = True)
-
-    # add new columns
-    navdata = navdata_1.concat(navdata_1)
-    assert navdata.shape == (4,12)
-    pandas_equiv = pd.concat((df_simple,df_simple),axis=0)
-    pandas_equiv.reset_index(drop=True, inplace=True)
-    pd.testing.assert_frame_equal(pandas_equiv.sort_index(axis=1),
-                                  navdata.pandas_df().sort_index(axis=1),
-                                  check_index_type=False,
-                                  check_dtype=False)
-
-    # add new rows
-    navdata = navdata_1.concat(navdata_1,axis=0)
-    assert navdata.shape == (8,6)
-    mapper = {"names":"names_0",
-              "floats":"floats_0",
-              "integers":"integers_0",
-              "strings":"strings_0"}
-    df_simple_2 = df_simple.rename(mapper,axis=1)
-    pandas_equiv = pd.concat((df_simple,df_simple_2),axis=1)
-    pandas_equiv.reset_index(drop=True, inplace=True)
-    pd.testing.assert_frame_equal(pandas_equiv.sort_index(axis=1),
-                                  navdata.pandas_df().sort_index(axis=1),
-                                  check_index_type=False,
-                                  check_dtype=False)
-
-    # concatenate empty NavData
-    navdata = NavData().concat(navdata_1,axis=1)
-    pd.testing.assert_frame_equal(df_simple.sort_index(axis=1),
-                                  navdata.pandas_df().sort_index(axis=1),
-                                  check_index_type=False,
-                                  check_dtype=False)
-    navdata = navdata_1.concat(NavData(),axis=1)
-    pd.testing.assert_frame_equal(df_simple.sort_index(axis=1),
-                                  navdata.pandas_df().sort_index(axis=1),
-                                  check_index_type=False,
-                                  check_dtype=False)
-
-    # test multiple rows with the same name
-    navdata_long = navdata_1.copy()
-    for count in range(13):
-        navdata_long.concat(navdata_1,axis=0,inplace=True)
-        for word in ["names","integers","floats","strings"]:
-            assert word + "_" + str(count) in navdata_long.rows
-
-    # add semi new columns
-    navdata = navdata_1.concat(navdata_2)
-    assert navdata.shape == (6,12)
-    assert np.all(np.isnan(navdata["floats"][-6:]))
-    assert np.all(navdata["names"][-6:] == np.array([np.nan]).astype(str)[0])
-    assert np.all(np.isnan(navdata["decimals"][:6]))
-    assert np.all(navdata["words"][:6] == np.array([np.nan]).astype(str)[0])
-
-    # add semi new columns in opposite order
-    navdata = navdata_2.concat(navdata_1)
-    assert navdata.shape == (6,12)
-    assert np.all(np.isnan(navdata["floats"][:6]))
-    assert np.all(navdata["names"][:6] == np.array([np.nan]).astype(str)[0])
-    assert np.all(np.isnan(navdata["decimals"][-6:]))
-    assert np.all(navdata["words"][-6:] == np.array([np.nan]).astype(str)[0])
-
-    # add as new rows
-    navdata = navdata_1.concat(navdata_2,axis=0)
-    assert navdata.shape == (8,6)
-    mapper = {"names":"words",
-              "floats":"decimals",
-              "integers":"integers_0",
-              "strings":"strings_0"}
-    df_simple_2 = df_simple.rename(mapper,axis=1)
-    pandas_equiv = pd.concat((df_simple,df_simple_2),axis=1)
-    pandas_equiv.reset_index(drop=True, inplace=True)
-    pd.testing.assert_frame_equal(pandas_equiv.sort_index(axis=1),
-                                  navdata.pandas_df().sort_index(axis=1),
-                                  check_index_type=False,
-                                  check_dtype=False)
-
-    navdata_a = NavData(pandas_df=pd.DataFrame({'a':[0],'b':[1],'c':[2],
-                                                'd':[3],'e':[4],'f':[5],
-                                                }))
-    navdata_b = navdata_a.concat(navdata_a.copy(),axis=0)
-    assert navdata_b.shape == (12,1)
-    navdata_b = navdata_a.concat(navdata_a.copy(),axis=1)
-    assert navdata_b.shape == (6,2)
-
-def test_concat_fails(df_simple):
-    """Test when concat should fail.
-
-    Parameters
-    ----------
-    df_simple : pd.DataFrame
-        Simple pd.DataFrame with which to initialize NavData.
-
-    """
-
-    navdata_1 = NavData(pandas_df=df_simple)
-
-    with pytest.raises(TypeError) as excinfo:
-        navdata_1.concat(np.array([]))
-    assert "concat" in str(excinfo.value)
-    assert "NavData" in str(excinfo.value)
-
-    navdata_2 = navdata_1.remove(cols=[0])
-
-    with pytest.raises(RuntimeError) as excinfo:
-        navdata_1.concat(navdata_2,axis=0)
-    assert "same length" in str(excinfo.value)
-    assert "concat" in str(excinfo.value)
-
-    with pytest.raises(RuntimeError) as excinfo:
-        navdata_1.concat(NavData(),axis=0)
-    assert "same length" in str(excinfo.value)
-    assert "concat" in str(excinfo.value)
-
-    with pytest.raises(RuntimeError) as excinfo:
-        NavData().concat(navdata_1,axis=0)
-    assert "same length" in str(excinfo.value)
-    assert "concat" in str(excinfo.value)
-
 @pytest.mark.parametrize("rows",
                         [None,
                         ['names', 'integers', 'floats', 'strings'],
@@ -2448,4 +2040,3 @@ def test_sort(data, df_simple):
     sorted_single_col = unsort_navdata_single_col.sort()
     pd.testing.assert_frame_equal(sorted_single_col.pandas_df(),
                                 unsort_navdata_single_col.pandas_df())
-
